Represent states as binary memory
This allows us represent all shift registers as memory variables where
we want to update them independently from rendering.  Want to allow us to write
as much and as efficiently as possible to write data to arduino memory,
but only write to shift registers when state dictates it.
Currently use state variables to store data and check for changes that
will trigger display/shift register refresh.